Move AtomicLinkedList into separate file
